moved getEvents to firebase/events.ts
